Fix calculator reset after equals sign pressed
Fixes a bug with the calculator equals sign. All globals get reset after an equation is calculated except for the inputWindow.value.

If the user selects a number after the equation the calculator will undergo a full clear.

If the user selects an operator they can continue doing maths with the result.